+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## [master](https://github.com/arangodb/kube-arangodb/tree/master) (N/A)
+- Add support for spec.ClusterDomain to be able to use FQDN in ArangoDB cluster communication
 
 ## [1.1.3](https://github.com/arangodb/kube-arangodb/tree/1.1.3) (2020-12-16)
 - Add v2alpha1 API for ArangoDeployment and ArangoDeploymentReplication